backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcore-stack"
version = "1.0.0-beta.1"
requires-python = ">=3.9"
description = "Multi-agent conversational AI system with AWS Bedrock AgentCore"
readme = "README.md"
license = {text = "MIT"}

# Core dependencies shared across both APIs
dependencies = [
    # FastAPI and web framework
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",

    # AWS and cloud services
    "boto3>=1.40.1",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "aiofiles>=25.1.0",

    # Authentication (for shared auth module)
    "pyjwt[crypto]>=2.8.0",

    # OAuth provider management
    "authlib>=1.3.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
# AgentCore-specific dependencies (for inference_api)
agentcore = [
    "strands-agents>=1.27.0",
    "strands-agents-tools==0.2.20",

    # "aws-opentelemetry-distro>=0.10.1",
    "bedrock-agentcore",
    # NOTE: nova-act temporarily disabled due to strands-agents version conflict
    # "nova-act==3.0.5.0",  # Requires strands-agents<=1.14.0
    # Multi-provider LLM support
    "openai>=1.0.0",  # For OpenAI models
    "google-genai>=1.0.0",  # For Google Gemini models
]

# Document ingestion pipeline dependencies (for Lambda deployment)
# NOTE: For document ingestion dependencies, see:
#   backend/src/apis/app_api/documents/ingestion/requirements.txt
# This is the single source of truth for the Lambda deployment.
# For local development: pip install -r backend/src/apis/app_api/documents/ingestion/requirements.txt

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-aiofiles>=25.1.0",
    "tiktoken>=0.5.0",
]

# All dependencies (convenience)
all = [
    "agentcore-stack[agentcore,dev]",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["agents*", "apis*"]

[tool.black]
line-length = 150
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 150
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
# pythonpath is relative to the directory where pytest is run (backend/)
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]
